+++ voltixApp/VoltixApp/Services/KeysignPayloadFactory.swift
@@ -6,6 +6,7 @@
 //
 
 import Foundation
+import BigInt
 
 struct KeysignPayloadFactory {
 
@@ -34,14 +35,14 @@ struct KeysignPayloadFactory {
     private let gaia = GaiaService.shared
     private let sol = SolanaService.shared
 
-    func buildTransfer(coin: Coin, toAddress: String, amount: Int64, memo: String?, chainSpecific: BlockChainSpecific, swapPayload: THORChainSwapPayload? = nil) async throws -> KeysignPayload {
+    func buildTransfer(coin: Coin, toAddress: String, amount: BigInt, memo: String?, chainSpecific: BlockChainSpecific, swapPayload: THORChainSwapPayload? = nil) async throws -> KeysignPayload {
 
         var utxos: [UtxoInfo] = []
 
         if case let .UTXO(byteFee) = chainSpecific {
-            let totalAmountNeeded = amount + byteFee
+            let totalAmountNeeded = amount + BigInt(byteFee)
 
-            guard let info = utxo.blockchairData[coin.blockchairKey]?.selectUTXOsForPayment(amountNeeded: totalAmountNeeded).map({
+            guard let info = utxo.blockchairData[coin.blockchairKey]?.selectUTXOsForPayment(amountNeeded: Int64(totalAmountNeeded)).map({
                 UtxoInfo(
                     hash: $0.transactionHash ?? "",
                     amount: Int64($0.value ?? 0),
@@ -56,7 +57,7 @@ struct KeysignPayloadFactory {
         return KeysignPayload(
             coin: coin,
             toAddress: toAddress,
-            toAmount: amount,
+            toAmount: BigInt(amount),
             chainSpecific: chainSpecific,
             utxos: utxos,
             memo: memo,
